requirements.txt
amqp==5.0.9
arabic-reshaper==2.1.3
asgiref==3.3.4
backcall==0.2.0
bcj-cffi==0.5.1
bcrypt==3.2.0
billiard==3.6.4.0
cached-property==1.5.2
celery==5.1.2
certifi==2020.12.5
cffi==1.14.5
chardet==4.0.0
charset-normalizer==2.0.3
click==7.1.2
click-didyoumean==0.0.3
click-plugins==1.1.1
click-repl==0.1.6
comtypes==1.1.10
coreapi==2.3.3
coreschema==0.0.4
coverage==4.5.4
cryptography==3.4.7
decorator==5.0.6
Django==3.2
django-auth-ldap==3.0.0
django-celery-beat==2.2.1
django-cors-headers==3.7.0
django-extensions==3.1.5
django-filter==2.4.0
django-rest-captcha==0.2.0
django-rosetta==0.9.6
django-timezone-field==4.2.3
djangorestframework==3.12.4
djangorestframework-simplejwt==4.4.0
drf-yasg==1.20.0
fabric==2.6.0
future==0.18.2
gunicorn==20.1.0
idna==2.10
importlib-metadata==3.10.0
inflection==0.5.1
invoke==1.5.0
ipdb==0.13.9
ipython==7.16.2
ipython-genutils==0.2.0
itypes==1.2.0
jedi==0.17.2
Jinja2==2.11.3
Khayyam==3.0.17
kombu==5.1.0
libvirt-python==7.10.0
lxml==4.7.1
MarkupSafe==1.1.1
multivolumefile==0.2.3
netifaces==0.11.0
numpy==1.19.5
packaging==20.9
pandas==1.1.5
paramiko==2.7.2
parso==0.7.1
pathlib2==2.3.5
pexpect==4.8.0
pickleshare==0.7.5
Pillow==8.2.0
plumbum==1.7.1
polib==1.1.1
prompt-toolkit==3.0.18
psutil==5.8.0
psycopg2-binary==2.8.6
ptyprocess==0.7.0
py7zr==0.15.2
pyasn1==0.4.8
pyasn1-modules==0.2.8
pycparser==2.20
pycryptodome==3.10.1
pydf==12
Pygments==2.8.1
PyJWT==1.7.1
PyNaCl==1.4.0
pyparsing==2.4.7
PyPDF2==1.26.0
pyppmd==0.12.1
pysftp==0.2.9
python-bidi==0.4.2
python-crontab==2.6.0
python-dateutil==2.8.2
python-dotenv==0.17.0
python-ldap==3.3.1
python-magic==0.4.22
pytz==2021.1
pyudev==0.22.0
pywinauto==0.6.6
PyYAML==6.0
pyzstd==0.14.4
rarfile==4.0
redis==3.5.3
reportlab==3.5.68
requests==2.25.1
rpyc==5.0.1
ruamel.yaml==0.17.4
ruamel.yaml.clib==0.2.2
six==1.15.0
sqlparse==0.4.1
texttable==1.6.3
toml==0.10.2
traitlets==4.3.3
typing-extensions==3.7.4.3
uritemplate==3.0.1
urllib3==1.26.4
vine==5.0.0
wcwidth==0.2.5
webdavclient3==3.14.6
websocket-client==1.1.0
zipp==3.4.1
